+++ simple-server.go
@@ -1,876 +1,18 @@
 package main
 
 import (
-	"crypto/rand"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log"
+	"2019_2_Solar/pkg/handls"
 	"net/http"
-	"os"
-	"regexp"
-	"strconv"
-	"sync"
-	"time"
 )
-
-var (
-	emailIsCorrect    = regexp.MustCompile(`^\w+@\w+\.\w+$`)
-	usernameIsCorrect = regexp.MustCompile(`^[\w\d]*$`)
-
-	passwordIsCorrect       = regexp.MustCompile(`^[\w\d!?_#&^%]{8,30}$`)
-	passwordHasDownCaseChar = regexp.MustCompile(`^.*[a-z]+.*$`)
-	passwordHasAperCaseChar = regexp.MustCompile(`^.*[A-Z]+.*$`)
-	passwordHasSpecChar     = regexp.MustCompile(`^.*[!?_#&^%]+.*$`)
-
-	nameIsCorrect    = regexp.MustCompile(`^[^\d_!@#$%^&*,.:~|\\\/\<\>=\+\?"'\[\]\{\}]*$`)
-	surnameIsCorrect = regexp.MustCompile(`^[^\d_!@#$%^&*,.:~|\\\/\<\>=\+\?"'\[\]\{\}]*$`)
-	statusIsCorrect  = regexp.MustCompile(`^.*$`)
-	ageIsCorrect     = regexp.MustCompile(`^[0-9]{1,3}$`)
-)
-
-type UserCookie struct {
-	Value      string    `json:"-"`
-	Expiration time.Time `json:"-"`
-}
-
-type UserSession struct {
-	ID     uint64 `json:"-"`
-	UserID uint64 `json:"-"`
-	UserCookie
-}
-
-type UserReg struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-	Username string `json:"username"`
-}
-
-type UserLogin struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-type EditUserProfile struct {
-	Username string `json:"username"`
-	Name     string `json:"name"`
-	Surname  string `json:"surname"`
-	Password string `json:"password"`
-	Email    string `json:"email"`
-	Age      string `json:"age"`
-	Status   string `json:"status"`
-	IsActive string `json:"isactive"`
-}
-
-type User struct {
-	ID        uint64 `json:"-"`
-	Username  string `json:"username"`
-	Name      string `json:"name"`
-	Surname   string `json:"surname"`
-	Password  string `json:"-"`
-	Email     string `json:"email"`
-	Age       string `json:"age"`
-	Status    string `json:"status"`
-	AvatarDir string `json:"-"`
-	IsActive  string `json:"isactive"`
-}
-
-type Handlers struct {
-	users    []User
-	sessions []UserSession
-	mu       *sync.Mutex
-}
-
-type DataJSON struct {
-	UserJSON  interface{} `json:"user,omitempty"`
-	UsersJSON interface{} `json:"users,omitempty"`
-	InfoJSON  interface{} `json:"info,omitempty"`
-}
-
-type OutJSON struct {
-	BodyJSON interface{} `json:"body"`
-}
-
-func CreateNewUser(h *Handlers, newUserReg UserReg) User {
-	var id uint64 = 0
-	if len(h.users) > 0 {
-		id = h.users[len(h.users)-1].ID + 1
-	}
-
-	newUser := User{
-		ID:       id,
-		Name:     "",
-		Password: newUserReg.Password,
-		Email:    newUserReg.Email,
-		Username: newUserReg.Username,
-	}
-	return newUser
-}
-
-func CreateNewUserSession(h *Handlers, user User) ([]http.Cookie, error) {
-	cookies := []http.Cookie{}
-
-	var sessionValue uint64 = 0
-	if len(h.sessions) > 0 {
-		sessionValue = h.sessions[len(h.sessions)-1].ID + 1
-	}
-
-	sessionKey := GenSessionKey(12)
-
-	cookieSessionKey := http.Cookie{
-		Name:    "session_key",
-		Value:   sessionKey,
-		Path:    "/",
-		Expires: time.Now().Add(1 * time.Hour),
-	}
-
-	cookies = append(cookies, cookieSessionKey)
-
-	newUserSession := UserSession{
-		ID:     sessionValue,
-		UserID: user.ID,
-		UserCookie: UserCookie{
-			Value:      sessionKey,
-			Expiration: time.Now().Add(1 * time.Hour),
-		},
-	}
-
-	h.sessions = append(h.sessions, newUserSession)
-
-	return cookies, nil
-}
-
-func DeleteOldUserSession(h *Handlers, value string) error {
-	for i, session := range h.sessions {
-		if session.Value == value {
-			h.sessions = append(h.sessions[:i], h.sessions[i+1:]...)
-			return nil
-		}
-	}
-	return errors.New("session has not found")
-}
-
-func SearchCookie(r *http.Request) (*http.Cookie, error) {
-	key, err := r.Cookie("session_key")
-	return key, err
-}
-
-func RegEmailIsUnique(h *Handlers, email string) bool {
-	for _, user := range h.users {
-		if user.Email == email {
-			return false
-		}
-	}
-	return true
-}
-
-func RegUsernameIsUnique(h *Handlers, username string) bool {
-	for _, user := range h.users {
-		if user.Username == username {
-			return false
-		}
-	}
-	return true
-}
-
-func EditEmailIsUnique(h *Handlers, email string, idUser uint64) bool {
-	for _, user := range h.users {
-		if user.Email == email {
-			if user.ID != idUser {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-func EditUsernameIsUnique(h *Handlers, username string, idUser uint64) bool {
-	for _, user := range h.users {
-		if user.Username == username {
-			if user.ID != idUser {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-func SearchUserByEmail(users []User, newUserLogin *UserLogin) interface{} {
-	for _, user := range users {
-		if user.Email == newUserLogin.Email {
-			return user
-		}
-	}
-	return ""
-}
-
-func GetUserIndexByID(h *Handlers, id uint64) int {
-	for index, user := range h.users {
-		if user.ID == id {
-			return index
-		}
-	}
-	return -1
-}
-
-func SetJsonData(data interface{}, infMsg string) OutJSON {
-	user, ok := data.(User)
-	if ok {
-		outJSON := OutJSON{
-			BodyJSON: DataJSON{
-				UserJSON: user,
-				InfoJSON: infMsg,
-			},
-		}
-		return outJSON
-	}
-	if users, ok := data.([]User); ok {
-
-		outJSON := OutJSON{
-			BodyJSON: DataJSON{
-				UsersJSON: users,
-				InfoJSON:  infMsg,
-			},
-		}
-		return outJSON
-	}
-	outJSON := OutJSON{
-		BodyJSON: DataJSON{
-			InfoJSON: infMsg,
-		},
-	}
-	return outJSON
-}
-
-func SearchIdUserByCookie(r *http.Request, h *Handlers) (uint64, error) {
-	sessionKey, err := SearchCookie(r)
-	if err == http.ErrNoCookie {
-		return 0, errors.New("cookies not found")
-	}
-
-	for _, oneSession := range h.sessions {
-		if oneSession.Value == sessionKey.Value {
-			return oneSession.UserID, nil
-		}
-	}
-	return 0, errors.New("idUser not found")
-}
-
-func SaveNewProfileUser(user *User, newUser *EditUserProfile) {
-	user.Age = newUser.Age
-	user.Status = newUser.Status
-	user.Name = newUser.Name
-	user.Surname = newUser.Surname
-
-	if newUser.Email != "" {
-		user.Email = newUser.Email
-	}
-	if newUser.Username != "" {
-		user.Username = newUser.Username
-	}
-	if newUser.Password != "" {
-		user.Password = newUser.Password
-	}
-}
-
-func ExtractFormatFile(FileName string) (string, error) {
-	for i := 0; i < len(FileName); i++ {
-		if string(FileName[i]) == "." {
-			return FileName[i:], nil
-		}
-	}
-	return "", errors.New("invalid file name")
-}
-
-func SetResponseError(encoder *json.Encoder, msg string, err error) {
-	log.Printf("%s: %s", msg, err)
-	data := SetJsonData(nil, msg)
-	encoder.Encode(data)
-}
-
-const (
-	letterBytes   = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789" // 62 possibilities
-	letterIdxBits = 6                                                                // 6 bits to represent 64 possibilities / indexes
-	letterIdxMask = 1<<letterIdxBits - 1                                             // All 1-bits, as many as letterIdxBits
-)
-
-func GenSessionKey(length int) string {
-	result := make([]byte, length)
-	bufferSize := int(float64(length) * 1.3)
-	for i, j, randomBytes := 0, 0, []byte{}; i < length; j++ {
-		if j%bufferSize == 0 {
-			randomBytes = SecureRandomBytes(bufferSize)
-		}
-		if idx := int(randomBytes[j%length] & letterIdxMask); idx < len(letterBytes) {
-			result[i] = letterBytes[idx]
-			i++
-		}
-	}
-
-	return string(result)
-}
-
-// SecureRandomBytes returns the requested number of bytes using crypto/rand
-func SecureRandomBytes(length int) []byte {
-	var randomBytes = make([]byte, length)
-	_, err := rand.Read(randomBytes)
-	if err != nil {
-		log.Fatal("Unable to generate random bytes")
-	}
-	return randomBytes
-}
-
-func UsernameCheck(username string) error {
-	if len(username) >= 1 && len(username) <= 30 && usernameIsCorrect.MatchString(username) {
-		return nil
-	}
-	return errors.New("Incorrect username")
-}
-
-func EmailCheck(email string) error {
-	if emailIsCorrect.MatchString(email) {
-		return nil
-	}
-	return errors.New("Incorrect email")
-}
-
-func PasswordCheck(password string) error {
-	if len(password) >= 8 && len(password) <= 30 && passwordIsCorrect.MatchString(password) {
-		if passwordHasAperCaseChar.MatchString(password) {
-			if passwordHasDownCaseChar.MatchString(password) {
-				if passwordHasSpecChar.MatchString(password) {
-					return nil
-				}
-				return errors.New("Password has not special symbol")
-			}
-			return errors.New("Password has not symbol in down case")
-		}
-		return errors.New("Password has not symbol in upper case")
-	}
-	return errors.New("Incorrect password")
-}
-
-func NameCheck(name string) error {
-	if len(name) >= 1 && len(name) <= 30 && nameIsCorrect.MatchString(name) {
-		return nil
-	}
-	return errors.New("Incorrct name")
-}
-
-func SurnameCheck(surname string) error {
-	if len(surname) >= 1 && len(surname) <= 30 && surnameIsCorrect.MatchString(surname) {
-		return nil
-	}
-	return errors.New("Incorrect surname")
-}
-
-func AgeCheck(age string) error {
-	if ageIsCorrect.MatchString(age) {
-		return nil
-	}
-	return errors.New("Incorrect age")
-}
-
-func StatusCheck(status string) error {
-	if len(status) >= 1 && len(status) <= 200 && statusIsCorrect.MatchString(status) {
-		return nil
-	}
-	return errors.New("Incorrect status")
-}
-
-func EditProfileDataCheck(newProfileUser *EditUserProfile) error {
-	if newProfileUser.Email != "" {
-		if err := EmailCheck(newProfileUser.Email); err != nil {
-			//SetResponseError(encoder, "incorrect Email", err)
-			return err
-		}
-	}
-	if newProfileUser.Username != "" {
-		if err := UsernameCheck(newProfileUser.Username); err != nil {
-			//SetResponseError(encoder, "incorrect Username", err)
-			return err
-		}
-	}
-	if newProfileUser.Password != "" {
-		if err := PasswordCheck(newProfileUser.Password); err != nil {
-			//SetResponseError(encoder, "incorrect Password", err)
-			return err
-		}
-	}
-	if newProfileUser.Name != "" {
-		if err := NameCheck(newProfileUser.Name); err != nil {
-			//SetResponseError(encoder, "incorrect Name", err)
-			return err
-		}
-	}
-	if newProfileUser.Surname != "" {
-		if err := SurnameCheck(newProfileUser.Surname); err != nil {
-			//SetResponseError(encoder, "incorrect Surname", err)
-			return err
-		}
-	}
-	if newProfileUser.Status != "" {
-		if err := StatusCheck(newProfileUser.Status); err != nil {
-			//SetResponseError(encoder, "incorrect Status", err)
-			return err
-		}
-	}
-	if newProfileUser.Age != "" {
-		if err := AgeCheck(newProfileUser.Age); err != nil {
-			//SetResponseError(encoder, "incorrect Status", err)
-			return err
-		}
-	}
-	return nil
-}
-
-func (h *Handlers) HandleEmpty(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-	data := SetJsonData(nil, "Empty handler has been done")
-	encoder.Encode(data)
-	log.Printf("Empty handler has been done")
-}
-
-func (h *Handlers) HandleRegUser(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-	decoder := json.NewDecoder(r.Body)
-
-	newUserReg := new(UserReg)
-	err := decoder.Decode(newUserReg)
-	if err != nil {
-		SetResponseError(encoder, "incorrect json", err)
-		return
-	}
-
-	if err := EmailCheck(newUserReg.Email); err != nil {
-		SetResponseError(encoder, "incorrect Email", err)
-		return
-	}
-	if err := UsernameCheck(newUserReg.Username); err != nil {
-		SetResponseError(encoder, "incorrect Username", err)
-		return
-	}
-	if err := PasswordCheck(newUserReg.Password); err != nil {
-		SetResponseError(encoder, err.Error(), err)
-		return
-	}
-
-	defer h.mu.Unlock()
-
-	h.mu.Lock()
-	if !RegEmailIsUnique(h, newUserReg.Email) {
-		SetResponseError(encoder, "not unique Email", errors.New("not unique Email"))
-		return
-	}
-	if !RegUsernameIsUnique(h, newUserReg.Username) {
-		SetResponseError(encoder, "not unique Username", errors.New("not unique Username"))
-		return
-	}
-
-	newUser := CreateNewUser(h, *newUserReg)
-	h.users = append(h.users, newUser)
-	cookies, err := CreateNewUserSession(h, newUser)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "error while generating sessionValue", err)
-		return
-	}
-	if len(cookies) < 1 {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "error while generating sessionValue", errors.New("incorrect while create session"))
-		return
-	}
-	http.SetCookie(w, &cookies[0])
-
-	data := SetJsonData(newUser, "OK")
-	err = encoder.Encode(data)
-	if err != nil {
-		SetResponseError(encoder, "bad user struct", err)
-		return
-	}
-}
-
-func (h *Handlers) HandleListUsers(w http.ResponseWriter, r *http.Request) {
-	encoder := json.NewEncoder(w)
-
-	h.mu.Lock()
-	data := SetJsonData(h.users, "OK")
-	h.mu.Unlock()
-
-	err := encoder.Encode(data)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "error while marshalling JSON", err)
-		return
-	}
-}
-
-func (h *Handlers) HandleLoginUser(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	decoder := json.NewDecoder(r.Body)
-	encoder := json.NewEncoder(w)
-
-	newUserLogin := new(UserLogin)
-	err := decoder.Decode(newUserLogin)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "incorrect json", err)
-		return
-	}
-
-	defer h.mu.Unlock()
-	h.mu.Lock()
-	value := SearchUserByEmail(h.users, newUserLogin)
-	user, ok := value.(User)
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "incorrect combination of Email and Password", errors.New("incorrect Email"))
-		return
-	}
-	if user.Password != newUserLogin.Password {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "incorrect combination of Email and Password", errors.New("incorrect Password"))
-		return
-	}
-
-	//Если пришли валидные куки, значит новую сессию не создаём
-	idUser, err := SearchIdUserByCookie(r, h)
-	fmt.Println(idUser)
-	if err == nil {
-		data := SetJsonData(user, "Successfully log in yet")
-		encoder.Encode(data)
-		return
-	}
-
-	cookies, err := CreateNewUserSession(h, user)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "error while generating sessionValue", err)
-		return
-	}
-	if len(cookies) < 1 {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "error while generating sessionValue", errors.New("incorrect while create session"))
-		return
-	}
-	http.SetCookie(w, &cookies[0])
-
-	data := SetJsonData(user, "OK")
-
-	err = encoder.Encode(data)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "bad user struct", err)
-		return
-	}
-}
-
-func (h *Handlers) HandleGetProfileUserData(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-	h.mu.Lock()
-	idUser, err := SearchIdUserByCookie(r, h)
-	h.mu.Unlock()
-	if err != nil {
-		w.WriteHeader(http.StatusUnauthorized)
-		SetResponseError(encoder, "invalid cookie or user", err)
-		return
-	}
-
-	h.mu.Lock()
-	data := SetJsonData(h.users[GetUserIndexByID(h, idUser)], "OK")
-	h.mu.Unlock()
-
-	err = encoder.Encode(data)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "bad user struct", err)
-		return
-	}
-}
-
-func (h *Handlers) HandleGetProfileUserPicture(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-
-	h.mu.Lock()
-	idUser, err := SearchIdUserByCookie(r, h)
-	h.mu.Unlock()
-
-	if err != nil {
-		w.WriteHeader(http.StatusUnauthorized)
-		SetResponseError(encoder, "invalid cookie or user", err)
-		return
-	}
-	h.mu.Lock()
-	filename := h.users[GetUserIndexByID(h, idUser)].AvatarDir
-	h.mu.Unlock()
-	openFile, err := os.Open(filename)
-	defer openFile.Close() //Close after function return
-	if err != nil {
-		//File not found, send 404
-		w.WriteHeader(http.StatusNotFound)
-		SetResponseError(encoder, "file not found", err)
-		return
-	}
-	//File is found, create and send the correct headers
-	//Get the Content-Type of the file
-	//Create a buffer to store the header of the file in
-	FileHeader := make([]byte, 512)
-	//Copy the headers into the FileHeader buffer
-	openFile.Read(FileHeader)
-	//Get content type of file
-	FileContentType := http.DetectContentType(FileHeader)
-
-	//Get the file size
-	FileStat, _ := openFile.Stat()                     //Get info from file
-	FileSize := strconv.FormatInt(FileStat.Size(), 10) //Get file size as a string
-
-	//Send the headers
-	w.Header().Set("Content-Disposition", "attachment; filename="+filename)
-	w.Header().Set("Content-Type", FileContentType)
-	w.Header().Set("Content-Length", FileSize)
-	//Send the file
-	//We read 512 bytes from the file already, so we reset the offset back to 0
-	openFile.Seek(0, 0)
-	io.Copy(w, openFile) //'Copy' the file to the client
-}
-
-func (h *Handlers) HandleEditProfileUserData(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	decoder := json.NewDecoder(r.Body)
-	encoder := json.NewEncoder(w)
-
-	newProfileUser := new(EditUserProfile)
-	err := decoder.Decode(newProfileUser)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "incorrect json", err)
-		return
-	}
-
-	if err := EditProfileDataCheck(newProfileUser); err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, err.Error(), err)
-		return
-	}
-
-	defer h.mu.Unlock()
-	h.mu.Lock()
-
-	idUser, err := SearchIdUserByCookie(r, h)
-	if err != nil {
-		w.WriteHeader(http.StatusUnauthorized)
-		SetResponseError(encoder, "invalid cookie or user", err)
-		return
-	}
-	if !EditEmailIsUnique(h, newProfileUser.Email, idUser) {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "not unique Email", errors.New("not unique Email"))
-		return
-	}
-	if !EditUsernameIsUnique(h, newProfileUser.Username, idUser) {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "not unique Username", errors.New("not unique Username"))
-		return
-	}
-
-	SaveNewProfileUser(&h.users[GetUserIndexByID(h, idUser)], newProfileUser)
-
-	data := SetJsonData(nil, "data successfully saved")
-	encoder.Encode(data)
-}
-
-func (h *Handlers) HandleLogoutUser(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-
-	sessionKey, err := SearchCookie(r)
-	if err == http.ErrNoCookie {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "Cookie has not found", err)
-		return
-	}
-
-	h.mu.Lock()
-	err = DeleteOldUserSession(h, sessionKey.Value)
-	h.mu.Unlock()
-
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "Session has not found", err)
-		return
-	}
-	sessionKey.Path = "/"
-	sessionKey.Expires = time.Now().AddDate(0, 0, -999)
-	http.SetCookie(w, sessionKey)
-
-	data := SetJsonData(nil, "Session has been successfully deleted")
-	encoder.Encode(data)
-}
-
-func (h *Handlers) HandleEditProfileUserPicture(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-
-	encoder := json.NewEncoder(w)
-	r.ParseMultipartForm(5 * 1024 * 1025)
-	h.mu.Lock()
-	idUser, err := SearchIdUserByCookie(r, h)
-	h.mu.Unlock()
-	if err != nil {
-		w.WriteHeader(http.StatusUnauthorized)
-		SetResponseError(encoder, "user not found or not valid cookies", err)
-		return
-	}
-	file, header, err := r.FormFile("profilePicture")
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "Cannot read profile picture", err)
-		return
-	}
-
-	defer file.Close()
-	formatFile, err := ExtractFormatFile(header.Filename)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "Cannot read profile picture", err)
-		return
-	}
-	fileName := strconv.FormatUint(idUser, 10) + "_picture" + formatFile
-	newFile, err := os.Create(fileName)
-	h.mu.Lock()
-	h.users[GetUserIndexByID(h, idUser)].AvatarDir = fileName
-	h.mu.Unlock()
-	defer newFile.Close()
-	_, err = io.Copy(newFile, file)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		SetResponseError(encoder, "File recording has failed", err)
-		return
-	}
-
-	data := SetJsonData(nil, "profile picture has been successfully saved")
-	encoder.Encode(data)
-}
-
-// ================================= Handler functions =================================
-
-func CORSMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-		w.Header().Set("Access-Control-Allow-Headers", "Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization")
-		w.Header().Set("Access-Control-Allow-Origin", "http://solar-env.v2zxh2s3me.us-east-2.elasticbeanstalk.com")
-		w.Header().Set("Access-Control-Allow-Credentials", "true")
-		next.ServeHTTP(w, r)
-	})
-}
-
-var handlers = Handlers{
-	users:    make([]User, 0),
-	sessions: make([]UserSession, 0),
-	mu:       &sync.Mutex{},
-}
-
-func HandleRoot(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.Write([]byte("{}"))
-}
-
-func HandleUsers(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	log.Println(r.URL.Path)
-	handlers.HandleListUsers(w, r)
-}
-
-func HandleRegistration(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleRegUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleLogin(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleLoginUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleLogout(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleLogoutUser(w, r)
-		return
-	}
-
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleProfileData(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleEditProfileUserData(w, r)
-		return
-	}
-	if r.Method == http.MethodGet {
-		handlers.HandleGetProfileUserData(w, r)
-		return
-	}
-	handlers.HandleEmpty(w, r)
-}
-
-func HandleProfilePicture(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	log.Println(r.URL.Path)
-
-	if r.Method == http.MethodPost {
-		handlers.HandleEditProfileUserPicture(w, r)
-		return
-	}
-	if r.Method == http.MethodGet {
-		handlers.HandleGetProfileUserPicture(w, r)
-		return
-	}
-	handlers.HandleEmpty(w, r)
-}
 
 func main() {
-	http.Handle("/", CORSMiddleware(http.HandlerFunc(HandleRoot)))
-	http.Handle("/users/", CORSMiddleware(http.HandlerFunc(HandleUsers)))
-	http.Handle("/registration/", CORSMiddleware(http.HandlerFunc(HandleRegistration)))
-	http.Handle("/login/", CORSMiddleware(http.HandlerFunc(HandleLogin)))
-	http.Handle("/logout/", CORSMiddleware(http.HandlerFunc(HandleLogout)))
-	http.Handle("/profile/data", CORSMiddleware(http.HandlerFunc(HandleProfileData)))
-	http.Handle("/profile/picture", CORSMiddleware(http.HandlerFunc(HandleProfilePicture)))
+	http.Handle("/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleRoot)))
+	http.Handle("/users/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleUsers)))
+	http.Handle("/registration/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleRegistration)))
+	http.Handle("/login/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleLogin)))
+	http.Handle("/logout/", handls.CORSMiddleware(http.HandlerFunc(handls.HandleLogout)))
+	http.Handle("/profile/data", handls.CORSMiddleware(http.HandlerFunc(handls.HandleProfileData)))
+	http.Handle("/profile/picture", handls.CORSMiddleware(http.HandlerFunc(handls.HandleProfilePicture)))
 
 	http.ListenAndServe(":8080", nil)
 }
